[Map]Extracted Mapview to another class.
